backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.1.0
pydantic==2.5.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Error Handling & Logging
structlog==23.2.0
python-json-logger==2.0.7

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# External APIs
httpx==0.25.2
aiohttp==3.9.1

# Utilities
pydantic-settings==2.1.0